requirements.txt
python-telegram-bot==21.5
mysqlclient>=2.2.0
python-dotenv>=1.0.0
